Passage de tous les bonusAttaque / bonusDefense en double, afin de pouvoir utiliser Infinity au lieu de MAX_VALUE, pour pouvoir sommer sans passer dans les négatifs.

+++ src/com/utbm/smallWorld/Element.java
@@ -12,7 +12,7 @@ public abstract class Element implements Comparable<Element> {
 	/**
 	 * Si retourne Integer.MAX_VALUE, alors inattaquable
 	 */
-	public int bonusDefense(Peuple attaquant) {
+	public double bonusDefense(Peuple attaquant) {
 		return 0;
 	}
